+++ 1.20/src/main/java/io/github/axolotlclient/modules/hud/gui/keystrokes/KeyBindsList.java
@@ -62,7 +62,7 @@ public class KeyBindsList extends ElementListWidget<KeyBindsList.Entry> {
 				this.maxNameWidth = i;
 			}
 
-			this.addEntry(new KeyEntry(keyMapping, component));
+			this.addEntry(new KeyEntry(keyMapping));
 		}
 
 		addEntry(new SpacerEntry());
@@ -116,7 +116,7 @@ public class KeyBindsList extends ElementListWidget<KeyBindsList.Entry> {
 		private final Text name;
 		private final ButtonWidget configureButton, removeButton;
 
-		KeyEntry(final KeystrokeHud.Keystroke key, final Text name) {
+		KeyEntry(final KeystrokeHud.Keystroke key) {
 			this.key = key;
 			this.name = key.getKey().getKeyName();
 			this.configureButton = ButtonWidget.builder(CONFIGURE_BUTTON_TITLE, button -> client.setScreen(new ConfigureKeyBindScreen(keyBindsScreen, keyBindsScreen.hud, key, false)))
@@ -163,37 +163,43 @@ public class KeyBindsList extends ElementListWidget<KeyBindsList.Entry> {
 
 	public class NewEntry extends Entry {
 
-		private final ButtonWidget addButton, addSpecialButton;
+		private final ButtonWidget addButton, addSpecialButton, addCustomButton;
 		private final KeystrokeHud.Keystroke key = keyBindsScreen.hud.newStroke();
 
 		public NewEntry() {
 			this.addButton = ButtonWidget.builder(Text.translatable("keystrokes.stroke.add"), button -> client.setScreen(new ConfigureKeyBindScreen(keyBindsScreen, keyBindsScreen.hud, key, true)))
-				.positionAndSize(0, 0, 150, 20)
+				.width(100)
 				.build();
 			this.addSpecialButton = ButtonWidget.builder(Text.translatable("keystrokes.stroke.add.special"),
 					button -> client.setScreen(new AddSpecialKeystrokeScreen(keyBindsScreen, keyBindsScreen.hud)))
-				.width(150).build();
+				.width(100).build();
+			this.addCustomButton = ButtonWidget.builder(Text.translatable("keystrokes.stroke.add.custom"),
+					button -> client.setScreen(new ConfigureKeyBindScreen(keyBindsScreen, keyBindsScreen.hud, keyBindsScreen.hud.newCustomStroke(), true)))
+				.width(100).build();
 		}
 
 		@Override
 		public List<? extends Selectable> selectableChildren() {
-			return List.of(addSpecialButton, addButton);
+			return List.of(addCustomButton, addSpecialButton, addButton);
 		}
 
 		@Override
 		public void render(GuiGraphics guiGraphics, int index, int top, int left, int width, int height, int mouseX, int mouseY, boolean hovering, float partialTick) {
-			int i = KeyBindsList.this.getScrollbarPositionX() - width / 2 - 10 + 4;
+			int i = KeyBindsList.this.getScrollbarPositionX() - width/2 - (300+8)/2 - 10;
 			int j = top - 2;
+			this.addCustomButton.setPosition(i, j);
+			this.addCustomButton.render(guiGraphics, mouseX, mouseY, partialTick);
+			i += addCustomButton.getWidth() + 4;
+			this.addSpecialButton.setPosition(i, j);
+			this.addSpecialButton.render(guiGraphics, mouseX, mouseY, partialTick);
+			i += addSpecialButton.getWidth() + 4;
 			this.addButton.setPosition(i, j);
 			this.addButton.render(guiGraphics, mouseX, mouseY, partialTick);
-			int k = i - addButton.getWidth() - 8;
-			this.addSpecialButton.setPosition(k, j);
-			this.addSpecialButton.render(guiGraphics, mouseX, mouseY, partialTick);
 		}
 
 		@Override
 		public List<? extends Element> children() {
-			return List.of(addSpecialButton, addButton);
+			return List.of(addCustomButton, addSpecialButton, addButton);
 		}
 	}
 }