+++ 1.20/src/main/java/io/github/axolotlclient/util/Util.java
@@ -41,9 +41,6 @@ import net.minecraft.util.Identifier;
 import org.apache.commons.lang3.StringUtils;
 
 public class Util {
-	public static String lastgame;
-	public static String game;
-
 	/**
 	 * Gets the amount of ticks in between start and end, on a 24000 tick system.
 	 *
@@ -112,7 +109,8 @@ public class Util {
 		Identifier id = new Identifier("axolotlclient", "graphics_" + name.toLowerCase(Locale.ROOT));
 		try {
 			NativeImageBackedTexture texture;
-			if (MinecraftClient.getInstance().getTextureManager().getOrDefault(id, null) == null) {
+			var previous = MinecraftClient.getInstance().getTextureManager().getOrDefault(id, null);
+			if (previous == null || (previous instanceof NativeImageBackedTexture tex && (tex.getImage().getHeight() != graphics.getHeight() || tex.getImage().getWidth() != graphics.getWidth()))) {
 				texture = new NativeImageBackedTexture(NativeImage.read(graphics.getPixelData()));
 				MinecraftClient.getInstance().getTextureManager().registerTexture(id, texture);
 			} else {
@@ -129,9 +127,5 @@ public class Util {
 			AxolotlClient.LOGGER.error("Failed to bind texture for " + name + ": ", e);
 		}
 		return id;
-	}
-
-	public static double lerp(double start, double end, double percent) {
-		return start + ((end - start) * percent);
 	}
 }