+++ 1.21/src/main/java/io/github/axolotlclient/util/Util.java
@@ -97,7 +97,8 @@ public class Util {
 		Identifier id = Identifier.of("axolotlclient", "graphics_" + name.toLowerCase(Locale.ROOT));
 		try {
 			NativeImageBackedTexture texture;
-			if (MinecraftClient.getInstance().getTextureManager().getOrDefault(id, null) == null) {
+			var previous = MinecraftClient.getInstance().getTextureManager().getOrDefault(id, null);
+			if (previous == null || (previous instanceof NativeImageBackedTexture tex && (tex.getImage().getHeight() != graphics.getHeight() || tex.getImage().getWidth() != graphics.getWidth()))) {
 				texture = new NativeImageBackedTexture(NativeImage.read(graphics.getPixelData()));
 				MinecraftClient.getInstance().getTextureManager().registerTexture(id, texture);
 			} else {
@@ -114,9 +115,5 @@ public class Util {
 			AxolotlClient.LOGGER.error("Failed to bind texture for " + name + ": ", e);
 		}
 		return id;
-	}
-
-	public static double lerp(double start, double end, double percent) {
-		return start + ((end - start) * percent);
 	}
 }