+++ 1.latest/src/main/java/io/github/axolotlclient/util/Util.java
@@ -96,10 +96,16 @@ public class Util {
 		Identifier id = Identifier.fromNamespaceAndPath("axolotlclient", "graphics_" + name.toLowerCase(Locale.ROOT));
 		try {
 			DynamicTexture texture;
-			if (!textures.containsKey(id)) {
+			boolean reuse = textures.containsKey(id);
+			if (reuse) {
+				var img = textures.get(id);
+				reuse = img.getPixels().getWidth() == graphics.getWidth() && img.getPixels().getHeight() == graphics.getHeight();
+			}
+			if (!reuse) {
 				texture = new DynamicTexture(id::toString, NativeImage.read(graphics.getPixelData()));
+				var prev = textures.put(id, texture);
+				if (prev != null) prev.close();
 				Minecraft.getInstance().getTextureManager().register(id, texture);
-				textures.put(id, texture);
 			} else {
 				texture = textures.get(id);
 				for (int x = 0; x < graphics.getWidth(); x++) {